Renderer changes; shader renamings, OpenGL texture binding to shaders changed
OpenGL texture binding to shaders is still a not solved problem. The
thing is that shaders are different and this is hard to code around.
One solution is to use the so called übershaders approach and ifdef
everything and then compile a unique shader for every unique setup..

This is something that must be solved in order to use different shader
and thus enable some fun shaderprogramming to happen! :D