EXT_texture_sRGB_decode used: sRGB conversion
Mipmapping needs to be done in linear texture space and if the texture,
as in the case of the diffuse textures, are in sRGB space the automatic
mipmapping performed by OpenGL will by incorrect (via the
glGenerateMipmap function). The solution is to use this broadly
supported extension such that the texture is always sampled as if it was
a linear RGB texture and as a side-effect the automatic mipmapping will
be correctly done in linear space (by encoding and decoding back and
forth).

Apparently this is super cheap for GPUs since they have hardware support
for sampling with sRGB/RGB so no perf. penalty is to be expected from this.